Wrote script to populate Locations table
populate_locations_table.py populates obtains the latitude and
longitude coordinates for every locations in our aggregated list,
and populates the Locations table with that data.

This has been tested with a few locations, but not yet run with
all of them. Need to make small changes (indicated in the script)
before running it "for real".